# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": { "title": "Growth OS API", "version": "0.1.0" },
  "paths": {
...
    "/api/v1/reviews": {
      "get": {
        "operationId": "listReviews",
        "responses": {
          "200": {
            "description": "Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": { "type": "string" },
                          "reviewerName": { "type": "string" },
                          "rating": { "type": "integer", "minimum": 1, "maximum": 5 },
                          "text": { "type": "string" },
                          "source": { "type": "string" },
                          "receivedAt": { "type": "string" },
                          "response": {
                            "oneOf": [
                              {
                                "type": "object",
                                "properties": {
                                  "kind": { "type": "string", "enum": ["unanswered"] }
                                },
                                "required": ["kind"],
                                "additionalProperties": false
                              },
                              {
                                "type": "object",
                                "properties": {
                                  "kind": { "type": "string", "enum": ["drafted"] },
                                  "text": { "type": "string" },
                                  "revisedAt": { "type": "string" }
                                },
                                "required": ["kind", "text", "revisedAt"],
                                "additionalProperties": false
                              },
                              {
                                "type": "object",
                                "properties": {
                                  "kind": { "type": "string", "enum": ["posted_manually"] },
                                  "text": { "type": "string" },
                                  "postedAt": { "type": "string" }
                                },
                                "required": ["kind", "text", "postedAt"],
                                "additionalProperties": false
                              }
                            ]
                          }
                        },
                        "required": [
                          "id",
                          "reviewerName",
                          "rating",
                          "text",
                          "source",
                          "receivedAt",
                          "response"
                        ],
                        "additionalProperties": false
                      }
                    },
                    "nextCursor": { "nullable": true, "type": "string" }
                  },
                  "required": ["items", "nextCursor"],
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "VALIDATION_FAILED",
                            "UNAUTHENTICATED",
                            "FORBIDDEN",
                            "RESOURCE_NOT_FOUND",
                            "INVALID_TRANSITION",
                            "IDEMPOTENCY_KEY_REUSED",
                            "RATE_LIMITED",
                            "CONFLICT",
                            "INTERNAL_ERROR",
                            "CSRF_FAILED",
                            "DUPLICATE_CUSTOMER",
                            "DUPLICATE_REVIEW_REQUIRED",
                            "CONSENT_REQUIRED",
                            "CAMPAIGN_VERSION_CONFLICT",
                            "CURRENCY_LOCKED",
                            "ACCOUNT_DELETION_REQUIRES_CONFIRMATION"
                          ]
                        },
                        "message": { "type": "string" },
                        "requestId": { "type": "string" },
                        "fieldErrors": {
                          "type": "object",
                          "additionalProperties": { "type": "array", "items": { "type": "string" } }
                        },
                        "details": { "type": "object", "additionalProperties": {} }
                      },
                      "required": ["code", "message", "requestId"],
                      "additionalProperties": false
                    }
                  },
                  "required": ["error"],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "createReview",
        "responses": {
          "201": {
            "description": "Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": { "type": "string" },
                    "reviewerName": { "type": "string" },
                    "rating": { "type": "integer", "minimum": 1, "maximum": 5 },
                    "text": { "type": "string" },
                    "source": { "type": "string" },
                    "receivedAt": { "type": "string" },
                    "response": {
                      "oneOf": [
                        {
                          "type": "object",
                          "properties": { "kind": { "type": "string", "enum": ["unanswered"] } },
                          "required": ["kind"],
                          "additionalProperties": false
                        },
                        {
                          "type": "object",
                          "properties": {
                            "kind": { "type": "string", "enum": ["drafted"] },
                            "text": { "type": "string" },
                            "revisedAt": { "type": "string" }
                          },
                          "required": ["kind", "text", "revisedAt"],
                          "additionalProperties": false
                        },
                        {
                          "type": "object",
                          "properties": {
                            "kind": { "type": "string", "enum": ["posted_manually"] },
                            "text": { "type": "string" },
                            "postedAt": { "type": "string" }
                          },
                          "required": ["kind", "text", "postedAt"],
                          "additionalProperties": false
                        }
                      ]
                    }
                  },
                  "required": [
                    "id",
                    "reviewerName",
                    "rating",
                    "text",
                    "source",
                    "receivedAt",
                    "response"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "VALIDATION_FAILED",
                            "UNAUTHENTICATED",
                            "FORBIDDEN",
                            "RESOURCE_NOT_FOUND",
                            "INVALID_TRANSITION",
                            "IDEMPOTENCY_KEY_REUSED",
                            "RATE_LIMITED",
                            "CONFLICT",
                            "INTERNAL_ERROR",
                            "CSRF_FAILED",
                            "DUPLICATE_CUSTOMER",
                            "DUPLICATE_REVIEW_REQUIRED",
                            "CONSENT_REQUIRED",
                            "CAMPAIGN_VERSION_CONFLICT",
                            "CURRENCY_LOCKED",
                            "ACCOUNT_DELETION_REQUIRES_CONFIRMATION"
                          ]
                        },
                        "message": { "type": "string" },
                        "requestId": { "type": "string" },
                        "fieldErrors": {
                          "type": "object",
                          "additionalProperties": { "type": "array", "items": { "type": "string" } }
                        },
                        "details": { "type": "object", "additionalProperties": {} }
                      },
                      "required": ["code", "message", "requestId"],
                      "additionalProperties": false
                    }
                  },
                  "required": ["error"],
                  "additionalProperties": false
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "reviewerName": { "type": "string", "minLength": 1, "maxLength": 120 },
                  "rating": {
                    "anyOf": [
                      { "type": "number", "enum": [1] },
                      { "type": "number", "enum": [2] },
                      { "type": "number", "enum": [3] },
                      { "type": "number", "enum": [4] },
                      { "type": "number", "enum": [5] }
                    ]
                  },
                  "text": { "type": "string", "minLength": 1, "maxLength": 5000 },
                  "source": { "type": "string", "minLength": 1, "maxLength": 80 },
                  "receivedAt": {
                    "type": "string",
                    "format": "date-time",
                    "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$"
                  }
                },
                "required": ["reviewerName", "rating", "text", "source"],
                "additionalProperties": false
              }
            }
          }
        }
      }
    },
    "/api/v1/reviews/{reviewId}/response": {
      "patch": {
        "operationId": "changeReviewResponse",
        "responses": {
          "200": {
            "description": "Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": { "type": "string" },
                    "reviewerName": { "type": "string" },
                    "rating": { "type": "integer", "minimum": 1, "maximum": 5 },
                    "text": { "type": "string" },
                    "source": { "type": "string" },
                    "receivedAt": { "type": "string" },
                    "response": {
                      "oneOf": [
                        {
                          "type": "object",
                          "properties": { "kind": { "type": "string", "enum": ["unanswered"] } },
                          "required": ["kind"],
                          "additionalProperties": false
                        },
                        {
                          "type": "object",
                          "properties": {
                            "kind": { "type": "string", "enum": ["drafted"] },
                            "text": { "type": "string" },
                            "revisedAt": { "type": "string" }
                          },
                          "required": ["kind", "text", "revisedAt"],
                          "additionalProperties": false
                        },
                        {
                          "type": "object",
                          "properties": {
                            "kind": { "type": "string", "enum": ["posted_manually"] },
                            "text": { "type": "string" },
                            "postedAt": { "type": "string" }
                          },
                          "required": ["kind", "text", "postedAt"],
                          "additionalProperties": false
                        }
                      ]
                    }
                  },
                  "required": [
                    "id",
                    "reviewerName",
                    "rating",
                    "text",
                    "source",
                    "receivedAt",
                    "response"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "VALIDATION_FAILED",
                            "UNAUTHENTICATED",
                            "FORBIDDEN",
                            "RESOURCE_NOT_FOUND",
                            "INVALID_TRANSITION",
                            "IDEMPOTENCY_KEY_REUSED",
                            "RATE_LIMITED",
                            "CONFLICT",
                            "INTERNAL_ERROR",
                            "CSRF_FAILED",
                            "DUPLICATE_CUSTOMER",
                            "DUPLICATE_REVIEW_REQUIRED",
                            "CONSENT_REQUIRED",
                            "CAMPAIGN_VERSION_CONFLICT",
                            "CURRENCY_LOCKED",
                            "ACCOUNT_DELETION_REQUIRES_CONFIRMATION"
                          ]
                        },
                        "message": { "type": "string" },
                        "requestId": { "type": "string" },
                        "fieldErrors": {
                          "type": "object",
                          "additionalProperties": { "type": "array", "items": { "type": "string" } }
                        },
                        "details": { "type": "object", "additionalProperties": {} }
                      },
                      "required": ["code", "message", "requestId"],
                      "additionalProperties": false
                    }
                  },
                  "required": ["error"],
                  "additionalProperties": false
                }
              }
            }
          },
          "404": {
            "description": "Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "VALIDATION_FAILED",
                            "UNAUTHENTICATED",
                            "FORBIDDEN",
                            "RESOURCE_NOT_FOUND",
                            "INVALID_TRANSITION",
                            "IDEMPOTENCY_KEY_REUSED",
                            "RATE_LIMITED",
                            "CONFLICT",
                            "INTERNAL_ERROR",
                            "CSRF_FAILED",
                            "DUPLICATE_CUSTOMER",
                            "DUPLICATE_REVIEW_REQUIRED",
                            "CONSENT_REQUIRED",
                            "CAMPAIGN_VERSION_CONFLICT",
                            "CURRENCY_LOCKED",
                            "ACCOUNT_DELETION_REQUIRES_CONFIRMATION"
                          ]
                        },
                        "message": { "type": "string" },
                        "requestId": { "type": "string" },
                        "fieldErrors": {
                          "type": "object",
                          "additionalProperties": { "type": "array", "items": { "type": "string" } }
                        },
                        "details": { "type": "object", "additionalProperties": {} }
                      },
                      "required": ["code", "message", "requestId"],
                      "additionalProperties": false
                    }
                  },
                  "required": ["error"],
                  "additionalProperties": false
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "type": "object",
                    "properties": {
                      "action": { "type": "string", "enum": ["save_draft"] },
                      "text": { "type": "string", "minLength": 1, "maxLength": 2000 }
                    },
                    "required": ["action", "text"],
                    "additionalProperties": false
                  },
                  {
                    "type": "object",
                    "properties": {
                      "action": { "type": "string", "enum": ["mark_posted_manually"] }
                    },
                    "required": ["action"],
                    "additionalProperties": false
                  },
                  {
                    "type": "object",
                    "properties": {
                      "action": { "type": "string", "enum": ["reopen_draft"] },
                      "text": { "type": "string", "minLength": 1, "maxLength": 2000 }
                    },
                    "required": ["action", "text"],
                    "additionalProperties": false
                  }
                ]
              }
            }
          }
        }
      }
    },
    "/api/v1/review-imports/preview": {
      "post": {
        "operationId": "previewReviewImport",
        "responses": {
          "200": {
            "description": "Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "importId": { "type": "string" },
                    "headers": { "type": "array", "items": { "type": "string" } },
                    "rows": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "rowNumber": {
                            "type": "integer",
                            "minimum": -9007199254740991,
                            "maximum": 9007199254740991
                          },
                          "values": {
                            "type": "object",
                            "additionalProperties": { "type": "string" }
                          },
                          "errors": { "type": "array", "items": { "type": "string" } }
                        },
                        "required": ["rowNumber", "values", "errors"],
                        "additionalProperties": false
                      }
                    },
                    "limits": {
                      "type": "object",
                      "properties": {
                        "maxRows": {
                          "type": "integer",
                          "minimum": -9007199254740991,
                          "maximum": 9007199254740991
                        },
                        "maxBytes": {
                          "type": "integer",
                          "minimum": -9007199254740991,
                          "maximum": 9007199254740991
                        }
                      },
                      "required": ["maxRows", "maxBytes"],
                      "additionalProperties": false
                    }
                  },
                  "required": ["importId", "headers", "rows", "limits"],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "VALIDATION_FAILED",
                            "UNAUTHENTICATED",
                            "FORBIDDEN",
                            "RESOURCE_NOT_FOUND",
                            "INVALID_TRANSITION",
                            "IDEMPOTENCY_KEY_REUSED",
                            "RATE_LIMITED",
                            "CONFLICT",
                            "INTERNAL_ERROR",
                            "CSRF_FAILED",
                            "DUPLICATE_CUSTOMER",
                            "DUPLICATE_REVIEW_REQUIRED",
                            "CONSENT_REQUIRED",
                            "CAMPAIGN_VERSION_CONFLICT",
                            "CURRENCY_LOCKED",
                            "ACCOUNT_DELETION_REQUIRES_CONFIRMATION"
                          ]
                        },
                        "message": { "type": "string" },
                        "requestId": { "type": "string" },
                        "fieldErrors": {
                          "type": "object",
                          "additionalProperties": { "type": "array", "items": { "type": "string" } }
                        },
                        "details": { "type": "object", "additionalProperties": {} }
                      },
                      "required": ["code", "message", "requestId"],
                      "additionalProperties": false
                    }
                  },
                  "required": ["error"],
                  "additionalProperties": false
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": { "csv": { "type": "string", "minLength": 1, "maxLength": 1000000 } },
                "required": ["csv"],
                "additionalProperties": false
              }
            }
          }
        }
      }
    },
...
  }
}